Created radix sort
Radix sort uses a subsorting (stable) algorithm of bucket sort. I implemented bucket sort in an unusual way. Instead of creating lists of buckets, I count the number of values in each bucket to determine their starting indeces. Then, I create an int array mapping the index in the sorted array to the index of the target value in the old array. Then, I iterate through that mapping and swap values in-place (possibly causing chains where I hop around to ensure that cyclic groups of index references are resolved). The benefit: the previous list is preserved by using the Collections.swap method, so that the implementation of the list remains. Is this better than creating bucket lists? Ehhhh, probably just on par. Still O(N) storage complexity.