Refactored some and started creating tensor kernels
Kernels are windows that can convolve over a tensor. Kernels provide a Tensor-like representations that is centered on a specific index of the target Tensor. This allow for something like a 3x3 kernel to convolve over a 128x128 tensor representing an image and apply some operation over that tensor using the kernel at each index (operations like Gaussian smoothing or LaPlacian transforms).